pyproject.toml
[project]
name = "ppt2desc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anthropic==0.42.0",
    "anyio==4.7.0",
    "boto3==1.35.91",
    "botocore==1.35.91",
    "cachetools==5.5.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "distro==1.9.0",
    "docstring-parser==0.16",
    "google-ai-generativelanguage==0.6.10",
    "google-api-core==2.24.0",
    "google-api-python-client==2.156.0",
    "google-auth==2.37.0",
    "google-auth-httplib2==0.2.0",
    "google-cloud-aiplatform==1.75.0",
    "google-cloud-bigquery==3.27.0",
    "google-cloud-core==2.4.1",
    "google-cloud-resource-manager==1.14.0",
    "google-cloud-storage==2.19.0",
    "google-crc32c==1.6.0",
    "google-genai==1.3.0",
    "google-generativeai==0.8.3",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.66.0",
    "grpc-google-iam-v1==0.13.1",
    "grpcio==1.68.1",
    "grpcio-status==1.68.1",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.8.2",
    "jmespath==1.0.1",
    "numpy==2.2.1",
    "openai==1.58.1",
    "packaging==24.2",
    "pillow==11.0.0",
    "proto-plus==1.25.0",
    "protobuf==5.29.2",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.1",
    "pydantic==2.10.4",
    "pydantic-core==2.27.2",
    "pymupdf==1.25.1",
    "pyparsing==3.2.1",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.3",
    "rsa==4.9",
    "s3transfer==0.10.4",
    "shapely==2.0.6",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tqdm==4.67.1",
    "typing-extensions==4.12.2",
    "uritemplate==4.1.1",
    "urllib3==2.3.0",
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
